exercise 4.21

+++ part4 - bloglist-app/controllers/blogs.js
@@ -4,7 +4,6 @@ const jwt = require('jsonwebtoken')
 const User = require("../models/user")
 const Blog = require("../models/blog")
 
-
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
     .find({}).populate('user', { username: 1, name: 1, id: 1})
@@ -35,6 +34,15 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const blog = await Blog.findById(request.params.id)
+  const user = await User.findById(decodedToken.id)
+  if(blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'invalid user' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   await response.status(204).end()
 })
